+++ opencl/Plotter.h
@@ -27,11 +27,11 @@
 class Plotter {
 public:
 
-    Plotter(unsigned int numNeurons,
-            unsigned int index,
+    Plotter(size_t numNeurons,
+            size_t index,
             float dt);
     void step(const state *curState,
-              const unsigned int t,
+              const size_t t,
               std::unique_ptr<float[]> const& sumFootprintAMPA,
               std::unique_ptr<float[]> const& sumFootprintNMDA,
               std::unique_ptr<float[]> const& sumFootprintGABAA);
@@ -39,8 +39,8 @@ public:
 
 private:
 
-    unsigned int _numNeurons;
-    unsigned int _index;
+    size_t _numNeurons;
+    size_t _index;
     float _dt;
 
     std::vector<float> _V, _h, _n, _z, _sAMPA, _xNMDA, _sNMDA, _IApp;
